Update modules to use renamed command module
