+++ src/discordBot/events/coursesChanged.js
@@ -1,4 +1,4 @@
-const { updateDynamicChoices } = require("../services/commands");
+const { updateDynamicChoices } = require("../services/command");
 
 const execute = async (event, client, Course) => {
   await updateDynamicChoices(client, ["join", "leave"], Course);
