+++ src/discordBot/events/interactionCreate.js
@@ -1,3 +1,5 @@
+const { sendErrorReport, sendErrorEphemeral } = require("../services/message");
+
 const execute = async (interaction, client, Course) => {
   if (!interaction.isCommand()) return;
   const command = client.slashCommands.get(interaction.commandName);
@@ -7,12 +9,12 @@ const execute = async (interaction, client, Course) => {
   }
   catch (error) {
     console.error(error);
-    await interaction.reply({ content: "There was an error while executing this command, report this problem to admins!", ephemeral: true });
+    await sendErrorReport(interaction, client, error);
+    await sendErrorEphemeral("There was an error while executing this command - Error report sent to administrators!");
   }
 };
 
 module.exports = {
   name: "interactionCreate",
-  ws: true,
   execute,
 };
